Removed buggy semi-colon from chat-room html file
When this html file was loaded, if there were messages already sent in the chat room, it worked fine, fine being all the old messages were displayed and the members of the room were displayed on the members area. However, if there was no messages sent it caused a syntax error. It was added by prettier formatter, which also added the brackets around the other jinja code but didn't add the brackets around this line.